split tools into two views
