+++ assets/src/components/workbenches/Workbenches.tsx
@@ -1,5 +1,4 @@
 import { useSetBreadcrumbs } from '@pluralsh/design-system'
-import { getTabCrumb } from 'components/ai/AI'
 import { SubTabs } from 'components/utils/SubTabs'
 import { useMemo } from 'react'
 import { Outlet, useMatch } from 'react-router-dom'
@@ -7,13 +6,16 @@ import {
   WORKBENCHES_ALERTS_REL_PATH,
   WORKBENCHES_ABS_PATH,
   WORKBENCHES_ISSUES_REL_PATH,
-  WORKBENCHES_TOOLS_REL_PATH,
+  WORKBENCHES_TOOLS_ADD_REL_PATH,
+  WORKBENCHES_TOOLS_YOUR_REL_PATH,
 } from 'routes/workbenchesRoutesConsts'
 import styled from 'styled-components'
+import { getTabCrumb } from 'components/ai/AI'
 
 const directory = [
   { label: 'Workbenches', path: '' },
-  { label: 'Tools', path: WORKBENCHES_TOOLS_REL_PATH },
+  { label: 'Add tools', path: WORKBENCHES_TOOLS_ADD_REL_PATH },
+  { label: 'Your created tools', path: WORKBENCHES_TOOLS_YOUR_REL_PATH },
   { label: 'Alerts', path: WORKBENCHES_ALERTS_REL_PATH },
   { label: 'Issues', path: WORKBENCHES_ISSUES_REL_PATH },
 ]
