+++ assets/src/components/workbenches/tools/WorkbenchToolsAdd.tsx
@@ -1,0 +1,103 @@
+import {
+  Button,
+  Card,
+  Chip,
+  Flex,
+  IconFrame,
+  AddIcon,
+} from '@pluralsh/design-system'
+import { CardGrid } from 'components/self-service/catalog/CatalogsGrid'
+import { StackedText } from 'components/utils/table/StackedText'
+import { Subtitle1H1 } from 'components/utils/typography/Text'
+import { WorkbenchTabHeader } from 'components/workbenches/common/WorkbenchTabHeader'
+import { WorkbenchTabWrapper } from 'components/workbenches/common/WorkbenchTabWrapper'
+import { useWorkbenchToolsQuery } from 'generated/graphql'
+import { isEmpty } from 'lodash'
+import { Link } from 'react-router-dom'
+import { WORKBENCHES_TOOLS_CREATE_ABS_PATH } from 'routes/workbenchesRoutesConsts'
+import styled, { useTheme } from 'styled-components'
+import { mapExistingNodes } from 'utils/graphql'
+import { TOOL_TYPE_CARDS, WorkbenchToolIcon } from './workbenchToolsUtils'
+
+const WORKBENCH_TOOL_TYPE_PARAM = 'type'
+
+export function WorkbenchToolsAdd() {
+  const { spacing } = useTheme()
+  const { data } = useWorkbenchToolsQuery()
+  const tools = mapExistingNodes(data?.workbenchTools)
+
+  return (
+    <WorkbenchTabWrapper>
+      <WorkbenchTabHeader
+        title="Integrations"
+        description="Setup and integrate your observability, infra, code, and custom tools using MCP, APIs, and webhooks natively with Plural."
+      />
+      <Subtitle1H1>Third party integrations</Subtitle1H1>
+      <CardGrid
+        styles={{
+          gridTemplateColumns: 'repeat(auto-fill, minmax(280px, 1fr))',
+        }}
+      >
+        {TOOL_TYPE_CARDS.map(({ type, description, label, categoryLabels }) => (
+          <ToolCardSC key={type}>
+            <StackedText
+              first={
+                <Flex
+                  align="center"
+                  gap="medium"
+                >
+                  <IconFrame
+                    circle
+                    type="secondary"
+                    icon={<WorkbenchToolIcon type={type} />}
+                  />
+                  {label}
+                </Flex>
+              }
+              firstPartialType="subtitle1"
+              firstColor="text"
+              second={isEmpty(tools) ? description : ''}
+              secondPartialType="body2"
+              secondColor="text-light"
+              gap="small"
+            />
+            <Flex
+              gap="xsmall"
+              wrap="wrap"
+              flex={1}
+            >
+              {categoryLabels.map((cat) => (
+                <Chip
+                  key={cat}
+                  size="small"
+                  css={{ height: 'fit-content' }}
+                >
+                  {cat}
+                </Chip>
+              ))}
+            </Flex>
+            <Button
+              small
+              floating
+              as={Link}
+              to={`${WORKBENCHES_TOOLS_CREATE_ABS_PATH}?${WORKBENCH_TOOL_TYPE_PARAM}=${type}`}
+              style={{ boxShadow: 'none', marginTop: spacing.xsmall }}
+              startIcon={<AddIcon />}
+            >
+              Add tool
+            </Button>
+          </ToolCardSC>
+        ))}
+      </CardGrid>
+    </WorkbenchTabWrapper>
+  )
+}
+
+const ToolCardSC = styled(Card)(({ theme }) => ({
+  display: 'flex',
+  flexDirection: 'column',
+  gap: theme.spacing.small,
+  padding: theme.spacing.large,
+  minHeight: 120,
+  textDecoration: 'none',
+}))
