+++ assets/src/routes/workbenchesRoutesConsts.tsx
@@ -15,9 +15,13 @@ export const WORKBENCHES_WEBHOOK_TRIGGERS_CREATE_WEBHOOK_REL_PATH =
 export const WORKBENCHES_WEBHOOK_SELECTED_QUERY_PARAM = 'selectedWebhook'
 
 export const WORKBENCHES_TOOLS_REL_PATH = 'tools'
+export const WORKBENCHES_TOOLS_ADD_REL_PATH = 'integrations'
+export const WORKBENCHES_TOOLS_YOUR_REL_PATH = 'your-integrations'
 export const WORKBENCHES_ALERTS_REL_PATH = 'alerts'
 export const WORKBENCHES_ISSUES_REL_PATH = 'issues'
 export const WORKBENCHES_TOOLS_ABS_PATH = `${WORKBENCHES_ABS_PATH}/${WORKBENCHES_TOOLS_REL_PATH}`
+export const WORKBENCHES_TOOLS_ADD_ABS_PATH = `${WORKBENCHES_ABS_PATH}/${WORKBENCHES_TOOLS_ADD_REL_PATH}`
+export const WORKBENCHES_TOOLS_YOUR_ABS_PATH = `${WORKBENCHES_ABS_PATH}/${WORKBENCHES_TOOLS_YOUR_REL_PATH}`
 export const WORKBENCHES_TOOLS_CREATE_ABS_PATH = `${WORKBENCHES_TOOLS_ABS_PATH}/${WORKBENCHES_CREATE_REL_PATH}`
 export const WORKBENCHES_TOOLS_EDIT_ABS_PATH = `${WORKBENCHES_TOOLS_ABS_PATH}/:${WORKBENCHES_TOOLS_PARAM_ID}/${WORKBENCHES_EDIT_REL_PATH}`
